Used type and offset to generate addresses for data segment entries
Converted tokens into entries. This gives parser easier access to the
data that it cares about while also making it easier to change the
specific details of the lexer.

This is a bit of a mess at the moment, but it's still growing. I don't
want to force a structure on it until I've tested a few things out.