Migrated the vm over to using the byte_code crate
There is an infinite loop, but I'm not sure if it's just because I got
the instruction encodings wrong. I'll investigate tomorrow.

Ended up restructuring the byte_code crate quite a lot. Which is what I
would expect from up front design. It's never the way to go. It's not
perfect, but it does the job for now.

Not every instruction is implemented in the VM yet, so there is still
some more work to do there. They're mostly the new ones that I thought of
while building out the byte_code crate, so it's fair. I'm not happy with
the `Operator` trait, but I can't think of a better way right now.