Added immediate versions of the control flow instructions
Now we have these:
call, callr, calli
jal,  jalr,  jali
jnz,  jnzr,  jnzi
jiz,  jizr,  jizi

The difference between the jump and call instructions is: Call works with
the stack and the frame pointer to keep track of all previous calls. Jump,
on the other hand, clobbers the link register so it's the responsiblity
of the programmer to save it if required. The trade off is that the jump
instructions are cheaper than the call instructions.

Also removed some debugging code from the encode tests.