Made some progress on the assembler!
Used code generation to generate parts of the lexer and parser. This is
driven by the `instr_codegen` crate, so changes there will propagate to
here :)

Almost implemented patching of label and data addresses. I need to do
some thinking in regards to the `loada` instruction. Currently we just
pack the address into the instruction, but that's limiting because we
can only fit an i16.

I also need to finish the actual patching. Currently we stop when we need
to replace the instruction. This is due to how Rust enums work. I don't
want to have a giant match statement just to fixup the handfull that
would be affected.

Finally, the abstraction layers have begun to leak into eachother. Which
tells me they were the wrong abstractions. I'll have to take a look at
this too.