Removed the the `size`, `f` or `s` fields in the instr_codegen crate
Noticed that no instructions use them and thus the code for them is
never generated. Those fields are only ever used during decoding.

So notice that despite changes to the data source and the build scripts
nothing changed in the generated code. Except in the one case of removing
the `parse_size` function.

This change has significantly simplified the encode/decode codegen :)