Removed the unsafe code in the convert module by remembering Rust's as cast rules
With integers the rules are:

Casting between integers of the same size is a no-op in both directions
i.e. `i64 -> u64 == u64 -> i64 == no-op`

Casting to a smaller size will truncate (i64 -> i8)
Casting to a larger size will:
 - Sign extend if the source is signed   `(i8 -> i64) || (i8 -> u64)`
 - Zero extend if the source is unsigned `(u8 -> i64) || (u8 -> u64)`

The float conversions still neeed to use the `to_bits`/`from_bits` methods,
so we still need the `Source` and `Destination` traits. The win is that
all the integer conversions can be unified.