Added a sign condition to i_type instructions
This lets the build script output `imm: u16` for those instructions. This
lets the vm do `imm as u64` without having to worry about accidentially
sign extending when it wasn't intended.

The downside is that the complexity of the build script has increased a
fair amount to avoid a exactly four of `as u16 as u64`, but it means if
the case comes up in the future or other devs refactor they're not going
to make the mistake.